Predict feature or not using LinearClassifier
This uses 3 GH features (additions, deletions, changed file count) and 2
velocity features (time to review, review cycles) to try to determine
whether a PR is a feature or not.

I wouldn't expect this to be a very good model, and it isn't - but, it is
better than random chance!

```
AUC on the validation set: 0.64
Accuracy on the validation set: 0.57
```